+++ virtualization/x86_vlapic/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Report the PIT timer and speaker-control ports as separate resource ranges so
+  unrelated host PIO is not claimed by the virtual PIT.
+
 ## [0.5.0](https://github.com/rcore-os/tgoskits/compare/x86_vlapic-v0.4.19...x86_vlapic-v0.5.0) - 2026-07-10
 
 ### Other